+++ app/controllers/api/discussions_controller.rb
@@ -1,15 +1,16 @@
 class API::DiscussionsController < API::RestfulController
+  include DashboardHelper
+
   load_and_authorize_resource only: [:show, :mark_as_read, :set_volume], find_by: :key
   load_resource only: [:create, :update]
 
   def dashboard_by_date
-    load_and_authorize_group if params[:group_id]
-    @discussions = page_collection inbox_threads
+    @discussions = page_collection dashboard_threads
     respond_with_discussions
   end
 
   def dashboard_by_group
-    @discussions = grouped inbox_threads.group_by(&:organization_id)
+    @discussions = grouped dashboard_threads.group_by(&:organization_id)
     respond_with_discussions
   end
 
@@ -59,19 +60,6 @@ class API::DiscussionsController < API::RestfulController
   end
 
   private
-
-  def inbox_threads
-    GroupDiscussionsViewer.for(user: current_user, group: @group, filter: params[:filter])
-                          .not_muted
-                          .where('last_activity_at > ?', 3.months.ago)
-                          .joined_to_current_motion
-                          .preload(:current_motion, {group: :parent})
-                          .order('motions.closing_at ASC, last_activity_at DESC')
-  end
-
-  def grouped(discussions)
-    discussions.map { |g, discussions| discussions.first(Integer(params[:per] || 5)) }.flatten
-  end
 
   def discussion_reader
     @dr ||= DiscussionReader.for(user: current_user, discussion: @discussion)
