+++ app/controllers/api/events_controller.rb
@@ -1,7 +1,13 @@
 class API::EventsController < API::RestfulController
+  include DashboardHelper
 
-  def dashboard_events
-    @events = DashboardEventsQuery.latest_events_for(user: current_user)
+  def dashboard_by_date
+    @events = page_collection dashboard_events
+    respond_with_collection
+  end
+
+  def dashboard_by_group
+    @events = grouped dashboard_events.group_by(&:organization_id)
     respond_with_collection
   end
