+++ lineman/app/models/event_records_interface.coffee
@@ -2,6 +2,12 @@ angular.module('loomioApp').factory 'EventRecordsInterface', (BaseRecordsInterfa
   class EventRecordsInterface extends BaseRecordsInterface
     model: EventModel
 
+    fetchDashboardByDate: (options = {}) ->
+      @restfulClient.get 'dashboard_by_date', options
+
+    fetchDashboardByGroup: (options = {}) ->
+      @restfulClient.get 'dashboard_by_group', options
+
     minLoadedSequenceIdByDiscussion: (discussion) ->
       item = _.min @where(discussionId: discussion.id), (event) -> event.sequenceId
       item.sequenceId || 0
