+++ spec/queries/latest_events_query_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-describe DashboardEventsQuery do
+describe LatestEventsQuery do
   let(:user) { create :user }
   let(:another_user) { create :user }
   let(:group) { create :group }
